add for features

+++ test.py
@@ -26,6 +26,10 @@ def main(cfg):
   if not os.path.exists(out_dir):
     os.makedirs(out_dir)
 
+  out_dir_features = os.path.join(cfg.test.out_dir, cfg.wandb.name, "features")
+  if not os.path.exists(out_dir_features):
+    os.makedirs(out_dir_features)
+
   postivs = list(Path(cfg.test.data_dir_pos).rglob('*.npy'))
   whole = list(Path(cfg.test.data_dir_whole).rglob('*.npy'))
   
@@ -34,14 +38,28 @@ def main(cfg):
   negs_txt = [x.strip().replace('.csv', '.npy') for x in negs_txt]
   negs = [x for x in whole if any([y in x.name for y in negs_txt])]
 
-  print(f"Positives: {len(postivs)}, Negatives: {len(negs)}")
   
-  for filepath in postivs + negs: 
+
+  if not cfg.test.whole:
+    files = postivs + negs
+    print(f"Positives: {len(postivs)}, Negatives: {len(negs)}")
+  else:
+    out_dir = os.path.join(cfg.test.out_dir, cfg.wandb.name, "csv_whole")
+    files = whole
+    print(f"Whole slides: {len(whole)}")
+  
+  for filepath in files: 
     dataset_id = filepath.name.split("_cleaned")[0].split(".npy")[0]
+
+    if os.path.exists(os.path.join(out_dir_features, f"{dataset_id}.npy")):
+      if os.path.exists(os.path.join(out_dir, f"{dataset_id}.csv")):
+        print(f"Skipping {dataset_id}")
+        continue
+    
     test_ds = SinglePatientDataset(filepath)
     test_loader = DataLoader(test_ds, batch_size=cfg.test.batch_size, shuffle=False, num_workers=cfg.test.num_workers)
 
-    probs, labels, _ ,_= get_outputs(model, test_loader)
+    probs, labels, _ ,features = get_outputs(model, test_loader)
     labels = labels.numpy()
 
     output_df = pd.DataFrame({
@@ -49,7 +67,25 @@ def main(cfg):
       'non-parasite output': probs[:,0],
       'parasite output': probs[:,1],
       'label': labels})
-    output_df.to_csv(os.path.join(out_dir, f"{dataset_id}.csv"), index=False)
+
+    if cfg.test.save_features:
+
+      # if already exists, skip
+      if os.path.exists(os.path.join(out_dir_features, f"{dataset_id}.npy")):
+        print(f"Skipping {dataset_id} features")
+      else:
+        # save to numpy array
+        print(f"Saving features for {dataset_id}")
+        features = features.numpy()
+        features = features.reshape(features.shape[0], -1)  # flatten
+        features.tofile(os.path.join(out_dir_features, f"{dataset_id}.npy"))
+
+        # check if the csv already exists
+    if os.path.exists(os.path.join(out_dir, f"{dataset_id}.csv")):
+      print(f"Skipping {dataset_id} predictions")
+      
+    else:
+      output_df.to_csv(os.path.join(out_dir, f"{dataset_id}.csv"), index=False)
 
 if __name__ == "__main__":
   main()
